Add method and demo for scanline intersection

+++ embedded-graphics/src/primitives/line/styled.rs
@@ -92,7 +92,7 @@ where
         if self.style.effective_stroke_color().is_some() {
             let (l, r) = self
                 .primitive
-                .extents(self.style.stroke_width, StrokeAlignment::Center);
+                .extents(self.style.stroke_width as i32, StrokeAlignment::Center);
 
             let min = l
                 .start
